+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 1.0.0 - 2024/01/16
 
-* TODO: Describe initial release.
+- add useDelayed hook for a delayed callback
+- add usePeriodic hook for a periodicly run callback
+- add useLoadingCallback hook for a triple state managed callback